% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scREALTIME.R
\name{getMetaData}
\alias{getMetaData}
\title{Input a Seurat object of single cell data and returns list object containing single cell trajectories}
\usage{
getMetaData(macro, stimulus, timepoints)
}
\arguments{
\item{macro}{Seurat object of single cell data with multiple measured timepoints}

\item{stimulus}{Stimulus name to label/subset the dataset}

\item{timepoints}{Vector of numeric timepoints in the dataset}
}
\description{
Input a Seurat object of single cell data and returns list object containing single cell trajectories
}
